--- src/mappings/core.ts
@@ -1,459 +1,0 @@
-import { BigDecimal, BigInt } from '@graphprotocol/graph-ts'
-
-import { Bundle, Burn, Factory, Mint, Pool, Swap, Tick, Token } from '../types/schema'
-import { Burn as BurnEvent, Initialize, Mint as MintEvent, Swap as SwapEvent } from '../types/templates/Pool/Pool'
-import { convertTokenToDecimal, loadTransaction, safeDiv } from '../utils'
-import { FACTORY_ADDRESS, ONE_BI, ZERO_BD } from '../utils/constants'
-import {
-  updatePoolDayData,
-  updatePoolHourData,
-  updateTokenDayData,
-  updateTokenHourData,
-  updateUniswapDayData,
-} from '../utils/intervalUpdates'
-import { findEthPerToken, getEthPriceInUSD, getTrackedAmountUSD, sqrtPriceX96ToTokenPrices } from '../utils/pricing'
-import { createTick } from '../utils/tick'
-
-export function handleInitialize(event: Initialize): void {
-  // update pool sqrt price and tick
-  const pool = Pool.load(event.address.toHexString())!
-  pool.sqrtPrice = event.params.sqrtPriceX96
-  pool.tick = BigInt.fromI32(event.params.tick)
-  pool.save()
-
-  // update token prices
-  const token0 = Token.load(pool.token0)
-  const token1 = Token.load(pool.token1)
-
-  // update ETH price now that prices could have changed
-  const bundle = Bundle.load('1')!
-  bundle.ethPriceUSD = getEthPriceInUSD()
-  bundle.save()
-
-  updatePoolDayData(event)
-  updatePoolHourData(event)
-
-  // update token prices
-  if (token0 && token1) {
-    token0.derivedETH = findEthPerToken(token0 as Token)
-    token1.derivedETH = findEthPerToken(token1 as Token)
-    token0.save()
-    token1.save()
-  }
-}
-
-export function handleMint(event: MintEvent): void {
-  const bundle = Bundle.load('1')!
-  const poolAddress = event.address.toHexString()
-  const pool = Pool.load(poolAddress)!
-  const factory = Factory.load(FACTORY_ADDRESS)!
-
-  const token0 = Token.load(pool.token0)
-  const token1 = Token.load(pool.token1)
-
-  if (token0 && token1) {
-    const amount0 = convertTokenToDecimal(event.params.amount0, token0.decimals)
-    const amount1 = convertTokenToDecimal(event.params.amount1, token1.decimals)
-
-    const amountUSD = amount0
-      .times(token0.derivedETH.times(bundle.ethPriceUSD))
-      .plus(amount1.times(token1.derivedETH.times(bundle.ethPriceUSD)))
-
-    // reset tvl aggregates until new amounts calculated
-    factory.totalValueLockedETH = factory.totalValueLockedETH.minus(pool.totalValueLockedETH)
-
-    // update globals
-    factory.txCount = factory.txCount.plus(ONE_BI)
-
-    // update token0 data
-    token0.txCount = token0.txCount.plus(ONE_BI)
-    token0.totalValueLocked = token0.totalValueLocked.plus(amount0)
-    token0.totalValueLockedUSD = token0.totalValueLocked.times(token0.derivedETH.times(bundle.ethPriceUSD))
-
-    // update token1 data
-    token1.txCount = token1.txCount.plus(ONE_BI)
-    token1.totalValueLocked = token1.totalValueLocked.plus(amount1)
-    token1.totalValueLockedUSD = token1.totalValueLocked.times(token1.derivedETH.times(bundle.ethPriceUSD))
-
-    // pool data
-    pool.txCount = pool.txCount.plus(ONE_BI)
-
-    // Pools liquidity tracks the currently active liquidity given pools current tick.
-    // We only want to update it on mint if the new position includes the current tick.
-    if (
-      pool.tick !== null &&
-      BigInt.fromI32(event.params.tickLower).le(pool.tick as BigInt) &&
-      BigInt.fromI32(event.params.tickUpper).gt(pool.tick as BigInt)
-    ) {
-      pool.liquidity = pool.liquidity.plus(event.params.amount)
-    }
-
-    pool.totalValueLockedToken0 = pool.totalValueLockedToken0.plus(amount0)
-    pool.totalValueLockedToken1 = pool.totalValueLockedToken1.plus(amount1)
-    pool.totalValueLockedETH = pool.totalValueLockedToken0
-      .times(token0.derivedETH)
-      .plus(pool.totalValueLockedToken1.times(token1.derivedETH))
-    pool.totalValueLockedUSD = pool.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    // reset aggregates with new amounts
-    factory.totalValueLockedETH = factory.totalValueLockedETH.plus(pool.totalValueLockedETH)
-    factory.totalValueLockedUSD = factory.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    const transaction = loadTransaction(event)
-    const mint = new Mint(transaction.id.toString() + '#' + pool.txCount.toString())
-    mint.transaction = transaction.id
-    mint.timestamp = transaction.timestamp
-    mint.pool = pool.id
-    mint.token0 = pool.token0
-    mint.token1 = pool.token1
-    mint.owner = event.params.owner
-    mint.sender = event.params.sender
-    mint.origin = event.transaction.from
-    mint.amount = event.params.amount
-    mint.amount0 = amount0
-    mint.amount1 = amount1
-    mint.amountUSD = amountUSD
-    mint.tickLower = BigInt.fromI32(event.params.tickLower)
-    mint.tickUpper = BigInt.fromI32(event.params.tickUpper)
-    mint.logIndex = event.logIndex
-
-    // tick entities
-    const lowerTickIdx = event.params.tickLower
-    const upperTickIdx = event.params.tickUpper
-
-    const lowerTickId = poolAddress + '#' + BigInt.fromI32(event.params.tickLower).toString()
-    const upperTickId = poolAddress + '#' + BigInt.fromI32(event.params.tickUpper).toString()
-
-    let lowerTick = Tick.load(lowerTickId)
-    let upperTick = Tick.load(upperTickId)
-
-    if (lowerTick === null) {
-      lowerTick = createTick(lowerTickId, lowerTickIdx, pool.id, event)
-    }
-
-    if (upperTick === null) {
-      upperTick = createTick(upperTickId, upperTickIdx, pool.id, event)
-    }
-
-    const amount = event.params.amount
-    lowerTick.liquidityGross = lowerTick.liquidityGross.plus(amount)
-    lowerTick.liquidityNet = lowerTick.liquidityNet.plus(amount)
-    upperTick.liquidityGross = upperTick.liquidityGross.plus(amount)
-    upperTick.liquidityNet = upperTick.liquidityNet.minus(amount)
-
-    lowerTick.save()
-    upperTick.save()
-
-    // TODO: Update Tick's volume, fees, and liquidity provider count. Computing these on the tick
-    // level requires reimplementing some of the swapping code from v3-core.
-
-    updateUniswapDayData(event)
-    updatePoolDayData(event)
-    updatePoolHourData(event)
-    updateTokenDayData(token0 as Token, event)
-    updateTokenDayData(token1 as Token, event)
-    updateTokenHourData(token0 as Token, event)
-    updateTokenHourData(token1 as Token, event)
-
-    token0.save()
-    token1.save()
-    pool.save()
-    factory.save()
-    mint.save()
-  }
-}
-
-export function handleBurn(event: BurnEvent): void {
-  const bundle = Bundle.load('1')!
-  const poolAddress = event.address.toHexString()
-  const pool = Pool.load(poolAddress)!
-  const factory = Factory.load(FACTORY_ADDRESS)!
-
-  const token0 = Token.load(pool.token0)
-  const token1 = Token.load(pool.token1)
-
-  if (token0 && token1) {
-    const amount0 = convertTokenToDecimal(event.params.amount0, token0.decimals)
-    const amount1 = convertTokenToDecimal(event.params.amount1, token1.decimals)
-
-    const amountUSD = amount0
-      .times(token0.derivedETH.times(bundle.ethPriceUSD))
-      .plus(amount1.times(token1.derivedETH.times(bundle.ethPriceUSD)))
-
-    // reset tvl aggregates until new amounts calculated
-    factory.totalValueLockedETH = factory.totalValueLockedETH.minus(pool.totalValueLockedETH)
-
-    // update globals
-    factory.txCount = factory.txCount.plus(ONE_BI)
-
-    // update token0 data
-    token0.txCount = token0.txCount.plus(ONE_BI)
-    token0.totalValueLocked = token0.totalValueLocked.minus(amount0)
-    token0.totalValueLockedUSD = token0.totalValueLocked.times(token0.derivedETH.times(bundle.ethPriceUSD))
-
-    // update token1 data
-    token1.txCount = token1.txCount.plus(ONE_BI)
-    token1.totalValueLocked = token1.totalValueLocked.minus(amount1)
-    token1.totalValueLockedUSD = token1.totalValueLocked.times(token1.derivedETH.times(bundle.ethPriceUSD))
-
-    // pool data
-    pool.txCount = pool.txCount.plus(ONE_BI)
-    // Pools liquidity tracks the currently active liquidity given pools current tick.
-    // We only want to update it on burn if the position being burnt includes the current tick.
-    if (
-      pool.tick !== null &&
-      BigInt.fromI32(event.params.tickLower).le(pool.tick as BigInt) &&
-      BigInt.fromI32(event.params.tickUpper).gt(pool.tick as BigInt)
-    ) {
-      pool.liquidity = pool.liquidity.minus(event.params.amount)
-    }
-
-    pool.totalValueLockedToken0 = pool.totalValueLockedToken0.minus(amount0)
-    pool.totalValueLockedToken1 = pool.totalValueLockedToken1.minus(amount1)
-    pool.totalValueLockedETH = pool.totalValueLockedToken0
-      .times(token0.derivedETH)
-      .plus(pool.totalValueLockedToken1.times(token1.derivedETH))
-    pool.totalValueLockedUSD = pool.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    // reset aggregates with new amounts
-    factory.totalValueLockedETH = factory.totalValueLockedETH.plus(pool.totalValueLockedETH)
-    factory.totalValueLockedUSD = factory.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    // burn entity
-    const transaction = loadTransaction(event)
-    const burn = new Burn(transaction.id + '#' + pool.txCount.toString())
-    burn.transaction = transaction.id
-    burn.timestamp = transaction.timestamp
-    burn.pool = pool.id
-    burn.token0 = pool.token0
-    burn.token1 = pool.token1
-    burn.owner = event.params.owner
-    burn.origin = event.transaction.from
-    burn.amount = event.params.amount
-    burn.amount0 = amount0
-    burn.amount1 = amount1
-    burn.amountUSD = amountUSD
-    burn.tickLower = BigInt.fromI32(event.params.tickLower)
-    burn.tickUpper = BigInt.fromI32(event.params.tickUpper)
-    burn.logIndex = event.logIndex
-
-    // tick entities
-    const lowerTickId = poolAddress + '#' + BigInt.fromI32(event.params.tickLower).toString()
-    const upperTickId = poolAddress + '#' + BigInt.fromI32(event.params.tickUpper).toString()
-    const lowerTick = Tick.load(lowerTickId)
-    const upperTick = Tick.load(upperTickId)
-    if (lowerTick && upperTick) {
-      const amount = event.params.amount
-      lowerTick.liquidityGross = lowerTick.liquidityGross.minus(amount)
-      lowerTick.liquidityNet = lowerTick.liquidityNet.minus(amount)
-      upperTick.liquidityGross = upperTick.liquidityGross.minus(amount)
-      upperTick.liquidityNet = upperTick.liquidityNet.plus(amount)
-
-      lowerTick.save()
-      upperTick.save()
-    }
-    updateUniswapDayData(event)
-    updatePoolDayData(event)
-    updatePoolHourData(event)
-    updateTokenDayData(token0 as Token, event)
-    updateTokenDayData(token1 as Token, event)
-    updateTokenHourData(token0 as Token, event)
-    updateTokenHourData(token1 as Token, event)
-
-    token0.save()
-    token1.save()
-    pool.save()
-    factory.save()
-    burn.save()
-  }
-}
-
-export function handleSwap(event: SwapEvent): void {
-  const bundle = Bundle.load('1')!
-  const factory = Factory.load(FACTORY_ADDRESS)!
-  const pool = Pool.load(event.address.toHexString())!
-
-  // hot fix for bad pricing
-  if (pool.id == '0x9663f2ca0454accad3e094448ea6f77443880454') {
-    return
-  }
-
-  const token0 = Token.load(pool.token0)
-  const token1 = Token.load(pool.token1)
-
-  if (token0 && token1) {
-    // amounts - 0/1 are token deltas: can be positive or negative
-    const amount0 = convertTokenToDecimal(event.params.amount0, token0.decimals)
-    const amount1 = convertTokenToDecimal(event.params.amount1, token1.decimals)
-
-    // need absolute amounts for volume
-    let amount0Abs = amount0
-    if (amount0.lt(ZERO_BD)) {
-      amount0Abs = amount0.times(BigDecimal.fromString('-1'))
-    }
-    let amount1Abs = amount1
-    if (amount1.lt(ZERO_BD)) {
-      amount1Abs = amount1.times(BigDecimal.fromString('-1'))
-    }
-
-    const amount0ETH = amount0Abs.times(token0.derivedETH)
-    const amount1ETH = amount1Abs.times(token1.derivedETH)
-    const amount0USD = amount0ETH.times(bundle.ethPriceUSD)
-    const amount1USD = amount1ETH.times(bundle.ethPriceUSD)
-
-    // get amount that should be tracked only - div 2 because cant count both input and output as volume
-    const amountTotalUSDTracked = getTrackedAmountUSD(amount0Abs, token0 as Token, amount1Abs, token1 as Token).div(
-      BigDecimal.fromString('2')
-    )
-    const amountTotalETHTracked = safeDiv(amountTotalUSDTracked, bundle.ethPriceUSD)
-    const amountTotalUSDUntracked = amount0USD.plus(amount1USD).div(BigDecimal.fromString('2'))
-
-    const feesETH = amountTotalETHTracked.times(pool.feeTier.toBigDecimal()).div(BigDecimal.fromString('1000000'))
-    const feesUSD = amountTotalUSDTracked.times(pool.feeTier.toBigDecimal()).div(BigDecimal.fromString('1000000'))
-
-    // global updates
-    factory.txCount = factory.txCount.plus(ONE_BI)
-    factory.totalVolumeETH = factory.totalVolumeETH.plus(amountTotalETHTracked)
-    factory.totalVolumeUSD = factory.totalVolumeUSD.plus(amountTotalUSDTracked)
-    factory.untrackedVolumeUSD = factory.untrackedVolumeUSD.plus(amountTotalUSDUntracked)
-    factory.totalFeesETH = factory.totalFeesETH.plus(feesETH)
-    factory.totalFeesUSD = factory.totalFeesUSD.plus(feesUSD)
-
-    // reset aggregate tvl before individual pool tvl updates
-    const currentPoolTvlETH = pool.totalValueLockedETH
-    factory.totalValueLockedETH = factory.totalValueLockedETH.minus(currentPoolTvlETH)
-
-    // pool volume
-    pool.volumeToken0 = pool.volumeToken0.plus(amount0Abs)
-    pool.volumeToken1 = pool.volumeToken1.plus(amount1Abs)
-    pool.volumeUSD = pool.volumeUSD.plus(amountTotalUSDTracked)
-    pool.untrackedVolumeUSD = pool.untrackedVolumeUSD.plus(amountTotalUSDUntracked)
-    pool.feesUSD = pool.feesUSD.plus(feesUSD)
-    pool.txCount = pool.txCount.plus(ONE_BI)
-
-    // Update the pool with the new active liquidity, price, and tick.
-    pool.liquidity = event.params.liquidity
-    pool.tick = BigInt.fromI32(event.params.tick as i32)
-    pool.sqrtPrice = event.params.sqrtPriceX96
-    pool.totalValueLockedToken0 = pool.totalValueLockedToken0.plus(amount0)
-    pool.totalValueLockedToken1 = pool.totalValueLockedToken1.plus(amount1)
-
-    // update token0 data
-    token0.volume = token0.volume.plus(amount0Abs)
-    token0.totalValueLocked = token0.totalValueLocked.plus(amount0)
-    token0.volumeUSD = token0.volumeUSD.plus(amountTotalUSDTracked)
-    token0.untrackedVolumeUSD = token0.untrackedVolumeUSD.plus(amountTotalUSDUntracked)
-    token0.feesUSD = token0.feesUSD.plus(feesUSD)
-    token0.txCount = token0.txCount.plus(ONE_BI)
-
-    // update token1 data
-    token1.volume = token1.volume.plus(amount1Abs)
-    token1.totalValueLocked = token1.totalValueLocked.plus(amount1)
-    token1.volumeUSD = token1.volumeUSD.plus(amountTotalUSDTracked)
-    token1.untrackedVolumeUSD = token1.untrackedVolumeUSD.plus(amountTotalUSDUntracked)
-    token1.feesUSD = token1.feesUSD.plus(feesUSD)
-    token1.txCount = token1.txCount.plus(ONE_BI)
-
-    // updated pool ratess
-    const prices = sqrtPriceX96ToTokenPrices(pool.sqrtPrice, token0 as Token, token1 as Token)
-    pool.token0Price = prices[0]
-    pool.token1Price = prices[1]
-    pool.save()
-
-    // update USD pricing
-    bundle.ethPriceUSD = getEthPriceInUSD()
-    bundle.save()
-    token0.derivedETH = findEthPerToken(token0 as Token)
-    token1.derivedETH = findEthPerToken(token1 as Token)
-
-    /**
-     * Things afffected by new USD rates
-     */
-    pool.totalValueLockedETH = pool.totalValueLockedToken0
-      .times(token0.derivedETH)
-      .plus(pool.totalValueLockedToken1.times(token1.derivedETH))
-    pool.totalValueLockedUSD = pool.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    factory.totalValueLockedETH = factory.totalValueLockedETH.plus(pool.totalValueLockedETH)
-    factory.totalValueLockedUSD = factory.totalValueLockedETH.times(bundle.ethPriceUSD)
-
-    token0.totalValueLockedUSD = token0.totalValueLocked.times(token0.derivedETH).times(bundle.ethPriceUSD)
-    token1.totalValueLockedUSD = token1.totalValueLocked.times(token1.derivedETH).times(bundle.ethPriceUSD)
-
-    // create Swap event
-    const transaction = loadTransaction(event)
-    const swap = new Swap(transaction.id + '#' + pool.txCount.toString())
-    swap.transaction = transaction.id
-    swap.timestamp = transaction.timestamp
-    swap.pool = pool.id
-    swap.token0 = pool.token0
-    swap.token1 = pool.token1
-    swap.sender = event.params.sender
-    swap.origin = event.transaction.from
-    swap.recipient = event.params.recipient
-    swap.amount0 = amount0
-    swap.amount1 = amount1
-    swap.amountUSD = amountTotalUSDTracked
-    swap.tick = BigInt.fromI32(event.params.tick as i32)
-    swap.sqrtPriceX96 = event.params.sqrtPriceX96
-    swap.logIndex = event.logIndex
-
-    // interval data
-    const uniswapDayData = updateUniswapDayData(event)
-    const poolDayData = updatePoolDayData(event)
-    const poolHourData = updatePoolHourData(event)
-    const token0DayData = updateTokenDayData(token0 as Token, event)
-    const token1DayData = updateTokenDayData(token1 as Token, event)
-    const token0HourData = updateTokenHourData(token0 as Token, event)
-    const token1HourData = updateTokenHourData(token1 as Token, event)
-
-    // update volume metrics
-    uniswapDayData.volumeETH = uniswapDayData.volumeETH.plus(amountTotalETHTracked)
-    uniswapDayData.volumeUSD = uniswapDayData.volumeUSD.plus(amountTotalUSDTracked)
-    uniswapDayData.feesUSD = uniswapDayData.feesUSD.plus(feesUSD)
-
-    poolDayData.volumeUSD = poolDayData.volumeUSD.plus(amountTotalUSDTracked)
-    poolDayData.volumeToken0 = poolDayData.volumeToken0.plus(amount0Abs)
-    poolDayData.volumeToken1 = poolDayData.volumeToken1.plus(amount1Abs)
-    poolDayData.feesUSD = poolDayData.feesUSD.plus(feesUSD)
-
-    poolHourData.volumeUSD = poolHourData.volumeUSD.plus(amountTotalUSDTracked)
-    poolHourData.volumeToken0 = poolHourData.volumeToken0.plus(amount0Abs)
-    poolHourData.volumeToken1 = poolHourData.volumeToken1.plus(amount1Abs)
-    poolHourData.feesUSD = poolHourData.feesUSD.plus(feesUSD)
-
-    token0DayData.volume = token0DayData.volume.plus(amount0Abs)
-    token0DayData.volumeUSD = token0DayData.volumeUSD.plus(amountTotalUSDTracked)
-    token0DayData.untrackedVolumeUSD = token0DayData.untrackedVolumeUSD.plus(amountTotalUSDTracked)
-    token0DayData.feesUSD = token0DayData.feesUSD.plus(feesUSD)
-
-    token0HourData.volume = token0HourData.volume.plus(amount0Abs)
-    token0HourData.volumeUSD = token0HourData.volumeUSD.plus(amountTotalUSDTracked)
-    token0HourData.untrackedVolumeUSD = token0HourData.untrackedVolumeUSD.plus(amountTotalUSDTracked)
-    token0HourData.feesUSD = token0HourData.feesUSD.plus(feesUSD)
-
-    token1DayData.volume = token1DayData.volume.plus(amount1Abs)
-    token1DayData.volumeUSD = token1DayData.volumeUSD.plus(amountTotalUSDTracked)
-    token1DayData.untrackedVolumeUSD = token1DayData.untrackedVolumeUSD.plus(amountTotalUSDTracked)
-    token1DayData.feesUSD = token1DayData.feesUSD.plus(feesUSD)
-
-    token1HourData.volume = token1HourData.volume.plus(amount1Abs)
-    token1HourData.volumeUSD = token1HourData.volumeUSD.plus(amountTotalUSDTracked)
-    token1HourData.untrackedVolumeUSD = token1HourData.untrackedVolumeUSD.plus(amountTotalUSDTracked)
-    token1HourData.feesUSD = token1HourData.feesUSD.plus(feesUSD)
-
-    swap.save()
-    token0DayData.save()
-    token1DayData.save()
-    uniswapDayData.save()
-    poolDayData.save()
-    poolHourData.save()
-    token0HourData.save()
-    token1HourData.save()
-    poolHourData.save()
-    factory.save()
-    pool.save()
-    token0.save()
-    token1.save()
-  }
-}
