+++ src/mappings/pool/index.ts
@@ -1,0 +1,6 @@
+import { handleBurn } from './burn'
+import { handleInitialize } from './initialize'
+import { handleMint } from './mint'
+import { handleSwap } from './swap'
+
+export { handleBurn, handleInitialize, handleMint, handleSwap }
